+++ tests/sqlalchemy/test_patch.py
@@ -1,10 +1,9 @@
 import datetime
-import json
 
 import pytest
 import sqlalchemy as sa
+from crate.client.http import json_dumps
 
-from cratedb_toolkit.util.sqlalchemy import CrateJsonEncoderWithNumPy
 from tests.conftest import TESTDRIVE_DATA_SCHEMA
 
 
@@ -48,18 +47,24 @@ def test_inspector_patched(database, needs_sqlalchemy2):
 def test_json_encoder_date():
     """
     Verify the extended JSON encoder also accepts Python's `date` types.
+
+    TODO: Move to different test file, as this no longer requires
+          monkeypatching after using orjson for marshalling.
     """
     data = {"date": datetime.date(2024, 6, 4)}
-    encoded = json.dumps(data, cls=CrateJsonEncoderWithNumPy)
-    assert encoded == '{"date": 1717459200000}'
+    encoded = json_dumps(data)
+    assert encoded == b'{"date":"2024-06-04"}'
 
 
 def test_json_encoder_numpy():
     """
     Verify the extended JSON encoder also accepts NumPy types.
+
+    TODO: Move to different test file, as this no longer requires
+          monkeypatching after using orjson for marshalling.
     """
     np = pytest.importorskip("numpy")
 
     data = {"scalar-int": np.float32(42.42).astype(int), "scalar-float": np.float32(42.42), "ndarray": np.ndarray([1])}
-    encoded = json.dumps(data, cls=CrateJsonEncoderWithNumPy)
-    assert encoded == """{"scalar-int": 42, "scalar-float": 42.41999816894531, "ndarray": [2.08e-322]}"""
+    encoded = json_dumps(data)
+    assert encoded == b"""{"scalar-int":42,"scalar-float":42.42,"ndarray":[2.08e-322]}"""