+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Fixed connectivity for `jobstats collect`
 - Refactored code and improved CLI interface of `ctk info` vs. `ctk cfr`
+- Dependencies: Updated to `crate-2.0.0`, which uses `orjson` for JSON marshalling
 
 ## 2025/01/13 v0.0.30
 - Dependencies: Minimize dependencies of core installation,